requirements.txt
django>=6.0
requests
python-dotenv
django-cors-headers
mysqlclient
djangorestframework
djangorestframework-simplejwt
